:hammer: refactor controls, settings menu, & widgets
