+++ packages/@ourworldindata/grapher/src/controls/LabeledSwitch.tsx
@@ -1,0 +1,45 @@
+import React from "react"
+import { observer } from "mobx-react"
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome/index.js"
+import { faInfoCircle } from "@fortawesome/free-solid-svg-icons"
+import { Tippy } from "@ourworldindata/utils"
+
+@observer
+export class LabeledSwitch extends React.Component<{
+    value?: boolean
+    label?: string
+    tooltip?: string
+    onToggle: () => any
+}> {
+    render(): JSX.Element {
+        const { label, value, tooltip } = this.props
+
+        return (
+            <div className="config-switch">
+                <label>
+                    <input
+                        type="checkbox"
+                        checked={value}
+                        onChange={this.props.onToggle}
+                    />
+                    <div className="outer">
+                        <div className="inner"></div>
+                    </div>
+                    {label}
+                    {tooltip && (
+                        <Tippy
+                            content={tooltip}
+                            theme="settings"
+                            placement="top"
+                            // arrow={false}
+                            maxWidth={338}
+                        >
+                            <FontAwesomeIcon icon={faInfoCircle} />
+                        </Tippy>
+                    )}
+                </label>
+                {tooltip && <div className="config-subtitle">{tooltip}</div>}
+            </div>
+        )
+    }
+}
