+++ packages/@ourworldindata/grapher/src/controls/settings/ZoomToggle.tsx
@@ -1,0 +1,30 @@
+import React from "react"
+import { action } from "mobx"
+import { observer } from "mobx-react"
+import { LabeledSwitch } from "../LabeledSwitch"
+
+export interface ZoomToggleManager {
+    zoomToSelection?: boolean
+}
+
+@observer
+export class ZoomToggle extends React.Component<{
+    manager: ZoomToggleManager
+}> {
+    @action.bound onToggle(): void {
+        this.props.manager.zoomToSelection = this.props.manager.zoomToSelection
+            ? undefined
+            : true
+    }
+
+    render(): JSX.Element {
+        return (
+            <LabeledSwitch
+                label="Zoom to selection"
+                tooltip="Scale axes to show only the currently highlighted data points."
+                value={this.props.manager.zoomToSelection}
+                onToggle={this.onToggle}
+            />
+        )
+    }
+}
