+++ packages/@ourworldindata/grapher/src/core/grapher.scss
@@ -60,6 +60,8 @@ $zindex-controls-drawer: 140;
 .GrapherComponent {
     @import "../controls/CommandPalette.scss";
     @import "../controls/Controls.scss";
+    @import "../controls/SettingsMenu.scss";
+    @import "../controls/MapProjectionMenu.scss";
     @import "../noDataModal/NoDataModal.scss";
     @import "../captionedChart/CaptionedChart.scss";
     @import "../timeline/TimelineComponent.scss";
